src/Chapter 7 - digital twins/azure-function-process-iot-adt/requirements.txt
# DO NOT include azure-functions-worker in this file
# The Python Worker is managed by Azure Functions platform
# Manually managing azure-functions-worker may cause unexpected issues

azure-functions==1.11.2
azure-digitaltwins-core==1.1.0
azure-identity==1.10.0